Test suite and fixed doomsday calculations
Implemented a basic test suite which allowed me to discover an error
when calculating the doomsday of a given year, which was then fixed.

Refactoring needs to be done throughout the program to make it more
readable and conform to C best practices.

More tests need to be added to check all variety of inputs and user
input needs to be verified that the date it valid. I am also unsure of
this point as how to treat years before 1582 (start of Gregorian
calender) when determining whether they year is a leap year or not.